# Conflicts:
#	Engine/Engine/Shaders/Source/Editor/SDFUI_ps.hlsl
#	Engine/Source/AssetSystemModule/Private/AssetSystem/AssetManager.cpp
#	Engine/Source/CoreUtilities/Public/CoreUtilities/Delegates/Delegate.h
#	Engine/Source/CoreUtilities/Public/CoreUtilities/Delegates/DelegateDeclarationHelpers.h
#	Engine/Source/CoreUtilities/Public/CoreUtilities/Delegates/DelegateHandle.h
#	Engine/Source/CoreUtilities/Public/CoreUtilities/Delegates/DelegateInstance.h
#	Engine/Source/ProjectUpgradeClient/Private/ProjectUpgradeClient/Upgrades/0_1_7/Serializers/FontSerializer.cpp
#	Engine/Source/Sandbox/Private/Sandbox/Sandbox.cpp
#	Engine/Source/Sandbox/Private/SandboxApp.cpp
#	Engine/Source/Sandbox/Public/Sandbox/Sandbox.h
#	Engine/Source/Sandbox/Public/Sandbox/Utility/EditorLibrary.h
#	Engine/Source/Sandbox/Sandbox.sharpmake.cs
#	Engine/Source/Volt-Assets/Private/Volt-Assets/Assets/Font.cpp
#	Engine/Source/Volt-Assets/Private/Volt-Assets/Font.cpp
#	Engine/Source/Volt-Assets/Private/Volt-Assets/MSDFData.h
#	Engine/Source/Volt-Assets/Volt-Assets.Sharpmake.cs
#	Engine/Source/Volt-Core/Public/Volt-Core/AssetTypes.h
#	Engine/Source/Volt/Private/Volt/Asset/Text/Font.cpp
#	Engine/Source/Volt/Private/Volt/Core/Application.cpp
#	Engine/Source/Volt/Public/Volt/Asset/Serializers/FontSerializer.h
#	Engine/Source/Volt/Volt.sharpmake.cs
#	Engine/Source/WindowModule/Private/WindowModule/Window.cpp

+++ Engine/Source/Volt-Renderer/Public/Volt-Renderer/RenderScene/SceneLightData.h
@@ -6,13 +6,15 @@
 
 #include <EntitySystem/EntityID.h>
 
+#include <CoreUtilities/UUID.h>
+
 #include <glm/glm.hpp>
 
 namespace Volt
 {
 	class RenderScene;
 
-	enum class SceneLightType : uint32_t
+	enum class SceneLightType : uint8_t
 	{
 		Directional = 0,
 		Point,
